Rename pkg/httptunnel -> pkg/ssh/sshtunnel
