+++ pkg/ssh/connector.go
@@ -62,19 +62,7 @@ func (c *Connector) Connect(host kubeoneapi.HostConfig) (Connection, error) {
 
 	conn, found := c.connections[host.ID]
 	if !found {
-		opts := Opts{
-			Username:    host.SSHUsername,
-			Port:        host.SSHPort,
-			Hostname:    host.PublicAddress,
-			KeyFile:     host.SSHPrivateKeyFile,
-			AgentSocket: host.SSHAgentSocket,
-			Timeout:     10 * time.Second,
-			Bastion:     host.Bastion,
-			BastionPort: host.BastionPort,
-			BastionUser: host.BastionUser,
-		}
-
-		conn, err = NewConnection(opts)
+		conn, err = NewConnection(c, sshOpts(host))
 		if err != nil {
 			return nil, err
 		}
@@ -83,4 +71,29 @@ func (c *Connector) Connect(host kubeoneapi.HostConfig) (Connection, error) {
 	}
 
 	return conn, nil
+}
+
+func (c *Connector) forgetConnection(conn *connection) {
+	c.lock.Lock()
+	defer c.lock.Unlock()
+
+	for k := range c.connections {
+		if c.connections[k] == conn {
+			delete(c.connections, k)
+		}
+	}
+}
+
+func sshOpts(host kubeoneapi.HostConfig) Opts {
+	return Opts{
+		Username:    host.SSHUsername,
+		Port:        host.SSHPort,
+		Hostname:    host.PublicAddress,
+		KeyFile:     host.SSHPrivateKeyFile,
+		AgentSocket: host.SSHAgentSocket,
+		Timeout:     10 * time.Second,
+		Bastion:     host.Bastion,
+		BastionPort: host.BastionPort,
+		BastionUser: host.BastionUser,
+	}
 }
